Don't reuse existing job when spawning a scheduled job
Since any job with non-inifinte result_ttl will be eventually cleaned up by RQ, it makes it impossible to use periodic jobs with finite or zero ttl. This change changes saved job into a template to store args, timeout and ttl, but actual executed job will be different so that the template will not be affected.