Fix bugs mentioned in issue 135.
This fixes enqueue_at and enqueue_in to pop out the things which _create_job
wants from the kwargs. That way we give _create_job the keyword args that
it will accept and leave the others to be passed along further.

There is a doctest in tests_issue_135.py to test this.