Only attempt to acquire a lock if there are jobs to enqueue.
There's no reason to acquire the lock if we have nothing to do.

This change is primarily useful to prevent the warning from being heavily spammed when multiple docker containers running redundant webservers all run a scheduler against the same queue. The warning will be printed only if one of the schedulers actually fails to enqueue a job due to another one having already acquired the lock.